Move import options to separate enum file
