+++ common/src/abstractions/import.service.ts
@@ -1,11 +1,8 @@
+import { ImportOption } from "../enums/importOptions";
 import { ImportError } from "../importers/importError";
 import { Importer } from "../importers/importer";
 import { ImportType } from "../services/import.service";
 
-export interface ImportOption {
-  id: string;
-  name: string;
-}
 export abstract class ImportService {
   featuredImportOptions: readonly ImportOption[];
   regularImportOptions: readonly ImportOption[];
